requirements.txt
fastapi==0.115.6
uvicorn[standard]~=0.34.0
pydantic~=2.10.4
requests
python-dotenv~=1.0.1
redis
motor~=3.6.0
pymongo~=4.9.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
httpx==0.24.1
pyotp==2.9.0
PyJWT~=2.10.1
starlette~=0.41.3
pydantic[email]
botocore~=1.35.97
boto3~=1.35.97
python-multipart
twilio~=9.6.0
asyncio~=3.4.3